Task3

+++ Tyuiu.YuzhakovYM.Sprint1.Task3.V17.Lib/DataService.cs
@@ -12,10 +12,10 @@ namespace Tyuiu.YuzhakovYM.Sprint1.Task3.V17.Lib
     {
         public bool ZeroCheck(double number)
         {
-            double number1, number2, number3;
-            number1 = (number * 10) % 10;
-            number2 = (number * 100) % 10;
-            number3 = (number * 1000) % 10;
+            int number1, number2, number3;
+            number1 = (int)(number * 10) % 10;
+            number2 = (int)(number * 100) % 10;
+            number3 = (int)(number * 1000) % 10;
             if (number1 == 0 || number2 == 0 || number3 == 0)
             {
                 return true;
